migrate object server tests
